move methods & update testing
