+++ tamr_unify_client/client.py
@@ -145,42 +145,6 @@ class Client:
         """
         return self._datasets
 
-    def create_project(self, project_creation_spec):
-        """
-        Create a Project in Unify
-
-        :param project_creation_spec: Project creation specification should be formatted as specified in the `Public Docs for Creating a Project <https://docs.tamr.com/reference#create-a-project>`_.
-        :type project_creation_spec: dict[str, str]
-        :returns: The created Project
-        :rtype: :class:`~tamr_unify_client.models.project.resource.Project`
-        """
-        from tamr_unify_client.models.project.resource import Project
-
-        data = (
-            self.post(self.projects.api_path, json=project_creation_spec)
-            .successful()
-            .json()
-        )
-        return Project.from_json(self, data)
-
-    def create_dataset(self, dataset_creation_spec):
-        """
-        Create a Dataset in Unify
-
-        :param dataset_creation_spec: Project creation specification should be formatted as specified in the `Public Docs for Creating a Dataset <https://docs.tamr.com/reference#create-dataset>`_.
-        :type dataset_creation_spec: dict[str, str]
-        :returns: The created Dataset
-        :rtype: :class:`~tamr_unify_client.models.dataset.resource.Dataset`
-        """
-        from tamr_unify_client.models.dataset.resource import Dataset
-
-        data = (
-            self.post(self.datasets.api_path, json=dataset_creation_spec)
-            .successful()
-            .json()
-        )
-        return Dataset.from_json(self, data)
-
     def __repr__(self):
         # Show only the type `auth` to mitigate any security concerns.
         return (
