+++ tamr_unify_client/models/project/collection.py
@@ -1,7 +1,6 @@
 from tamr_unify_client.models.base_collection import BaseCollection
 from tamr_unify_client.models.project.resource import Project
 
-
 class ProjectCollection(BaseCollection):
     """Collection of :class:`~tamr_unify_client.models.project.resource.Project` s.
 
@@ -61,5 +60,21 @@ class ProjectCollection(BaseCollection):
             >>>     do_stuff(project)
         """
         return super().stream(Project)
+ 
+    def create_project(self, project_creation_spec):
+        """
+        Create a Project in Unify
+
+        :param project_creation_spec: Project creation specification should be formatted as specified in the `Public Docs for Creating a Project <https://docs.tamr.com/reference#create-a-project>`_.
+        :type project_creation_spec: dict[str, str]
+        :returns: The created Project
+        :rtype: :class:`~tamr_unify_client.models.project.resource.Project`
+        """
+        data = (
+            self.client.post(self.api_path, json=project_creation_spec)
+            .successful()
+            .json()
+        )
+        return Project.from_json(self, data)
 
     # super.__repr__ is sufficient
